+++ packages/interaction-viewer/src/__test__/apiLoaderTest.spec.js
@@ -7,6 +7,7 @@ describe("interaction-viewer apiLoader", () => {
   it("should process the data", () => {
     const { data, diseases } = process(jsonData);
     // subcellulartreeMenu: TypeError: Converting circular structure to JSON
+    // console.log(JSON.stringify(data));
     expect(diseases).toEqual(diseasesJson);
     expect(data).toEqual(dataJson);
   });
