+++ packages/interaction-viewer/src/__test__/resources/diseases.json
@@ -1,23 +1,1 @@
-{
-  "Seckel syndrome 6": { "name": "Seckel syndrome 6", "selected": false },
-  "Duchenne muscular dystrophy": {
-    "name": "Duchenne muscular dystrophy",
-    "selected": false
-  },
-  "Becker muscular dystrophy": {
-    "name": "Becker muscular dystrophy",
-    "selected": false
-  },
-  "Cardiomyopathy, dilated, X-linked 3B": {
-    "name": "Cardiomyopathy, dilated, X-linked 3B",
-    "selected": false
-  },
-  "Hermansky-Pudlak syndrome 7": {
-    "name": "Hermansky-Pudlak syndrome 7",
-    "selected": false
-  },
-  "Infantile striatonigral degeneration": {
-    "name": "Infantile striatonigral degeneration",
-    "selected": false
-  }
-}
+[{ "name": "Seckel syndrome 6", "selected": false }]
